Cleanup and encapsulate response handling
