+++ mnt/influxdb/include/query_response_time.h
@@ -1,0 +1,8 @@
+#ifndef __DRIVER_MNT_INFLUXDB_QUERY_RESPONSE_TIME_H__
+#define __DRIVER_MNT_INFLUXDB_QUERY_RESPONSE_TIME_H__
+
+#include <driver/mnt/influxdb/influxdb.h>
+
+int16_t influxdb_Mount_response_time(JSON_Object *output, JSON_Value *value);
+
+#endif //__DRIVER_MNT_INFLUXDB_QUERY_RESPONSE_TIME_H__
